day11/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-mcp-adapters
faiss-cpu
pypdf
python-dotenv
sentence-transformers
mcp
fastmcp
langchain-google-genai
